lab2/requirements.txt
requests>=2.31.0
pymongo>=4.7.0
PyYAML>=6.0.1
